# api-quiz/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/quiz/users/{userId}": {
      "get": {
        "operationId": "readPublicUserInfo",
        "summary": "",
        "description": "Read public user information",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "updateUser",
        "summary": "",
        "description": "Update User",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUsersInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/quiz/users": {
      "get": {
        "operationId": "indexUsers",
        "summary": "",
        "description": "Index Users",
        "parameters": [
          {
            "name": "members",
            "required": false,
            "in": "query",
            "description": "Members",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "sortPointDesc",
            "required": false,
            "in": "query",
            "description": "Sort point desc",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isAdmin",
            "required": false,
            "in": "query",
            "description": "Include manager account",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UsersResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/quiz/admin/users": {
      "get": {
        "operationId": "indexUsers",
        "summary": "",
        "description": "Index Users",
        "parameters": [
          {
            "name": "members",
            "required": false,
            "in": "query",
            "description": "Members",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "sortPointDesc",
            "required": false,
            "in": "query",
            "description": "Sort point desc",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "isAdmin",
            "required": false,
            "in": "query",
            "description": "Include manager account",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "perPage",
            "required": false,
            "in": "query",
            "description": "Items per page",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UsersResponse"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin.users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "post": {
        "operationId": "createUser",
        "summary": "",
        "description": "Create User",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin.users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/api/quiz/admin/users/{userId}": {
      "get": {
        "operationId": "readUser",
        "summary": "",
        "description": "Read User",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin.users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "put": {
        "operationId": "updateUser",
        "summary": "",
        "description": "Update User",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUsersInputAdmin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin.users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      },
      "delete": {
        "operationId": "deleteUser",
        "summary": "",
        "description": "Delete User",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsersResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "503": {
            "description": "Service unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "tags": [
          "admin.users"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Quiz API",
    "description": "Swagger specification for api-quiz | [swagger.json](swagger.json)",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateUserInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "First and last name",
            "example": "Nguyễn Văn A"
          },
          "userName": {
            "type": "string",
            "description": "User name",
            "example": "veryvery"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Phone number",
            "example": "0981248920"
          },
          "email": {
            "type": "string",
            "description": "Email",
            "example": "example@email.com"
          },
          "imageUrl": {
            "type": "string",
            "description": "Avatar",
            "example": "https://www.studyphim.vn/system/movies/486/thumbnails/medium/Untitled.png"
          },
          "password": {
            "type": "string",
            "description": "Password",
            "example": "veryverysecret"
          }
        },
        "required": [
          "name",
          "userName",
          "phoneNumber",
          "email",
          "password"
        ]
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "number"
          },
          "message": {
            "type": "string",
            "example": "Error message",
            "default": "Internal Server Error"
          },
          "name": {
            "type": "string",
            "example": "Error"
          }
        }
      },
...
      "UsersResponse": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "First and last name",
            "example": "Nguyễn Văn A"
          },
          "userName": {
            "type": "string",
            "description": "User name",
            "example": "veryvery"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Phone number",
            "example": "0981248920"
          },
          "email": {
            "type": "string",
            "description": "Email",
            "example": "example@email.com"
          },
          "imageUrl": {
            "type": "string",
            "description": "Avatar",
            "example": "https://www.studyphim.vn/system/movies/486/thumbnails/medium/Untitled.png"
          },
          "code": {
            "type": "string",
            "description": "Code",
            "example": "MAGIOITHIEU06"
          },
          "status": {
            "type": "string",
            "description": "Status",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "DRAFT"
            ],
            "example": "DRAFT"
          },
          "gender": {
            "type": "string",
            "description": "Gender",
            "enum": [
              "MALE",
              "FEMALE",
              "OTHER"
            ],
            "example": "FEMALE"
          },
          "age": {
            "type": "number",
            "description": "Age",
            "example": 18
          },
          "target": {
            "type": "string",
            "description": "Target",
            "example": "Đỗ đại học"
          },
          "totalPoint": {
            "type": "number",
            "description": "totalPoint",
            "example": 862
          },
          "totalHeart": {
            "type": "number",
            "description": "totalHeart",
            "example": 235
          },
          "totalExam": {
            "type": "number",
            "description": "totalExam",
            "example": 3
          },
          "totalQuestions": {
            "type": "number",
            "description": "totalQuestions",
            "example": 62
          },
          "totalAnswers": {
            "type": "number",
            "description": "totalAnswers",
            "example": 16
          }
        },
        "required": [
          "name",
          "userName",
          "phoneNumber",
          "email"
        ]
      },
      "UpdateUsersInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name",
            "example": "Nguyễn Văn A"
          },
          "userName": {
            "type": "string",
            "description": "User name",
            "example": "veryvery"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Phone number",
            "example": "0981248920"
          },
          "email": {
            "type": "string",
            "description": "Email",
            "example": "example@email.com"
          },
          "imageUrl": {
            "type": "string",
            "description": "Avatar",
            "example": "https://www.studyphim.vn/system/movies/486/thumbnails/medium/Untitled.png"
          },
          "password": {
            "type": "string",
            "description": "Password",
            "example": "veryverysecret"
          }
        }
      },
      "UpdateUsersInputAdmin": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name",
            "example": "Nguyễn Văn A"
          },
          "userName": {
            "type": "string",
            "description": "User name",
            "example": "veryvery"
          },
          "phoneNumber": {
            "type": "string",
            "description": "Phone number",
            "example": "0981248920"
          },
          "email": {
            "type": "string",
            "description": "Email",
            "example": "example@email.com"
          },
          "imageUrl": {
            "type": "string",
            "description": "Avatar",
            "example": "https://www.studyphim.vn/system/movies/486/thumbnails/medium/Untitled.png"
          },
          "password": {
            "type": "string",
            "description": "Password",
            "example": "veryverysecret"
          },
          "status": {
            "type": "string",
            "description": "Status",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "DRAFT"
            ],
            "example": "DRAFT"
          }
        }
      },
...
    }
  }
}